Renaming & little changes
Ticket: #254
