+++ src/app/Main/Body/QueryPage/Results/TableResults/TableResults.model.ts
